backend/requirements.txt
fastapi[all]
python-dotenv
psycopg2-binary
pandas
fuzzywuzzy